+++ implementation/src/solution/optimal_partition_penalization.py
